Physics plugin now has basic functionality
This is a pretty massive overhaul of the physics plugin and it is by
no means ready to be merged yet. The code needs a lot of cleanup and
love in order to be brought up to standards with the rest of the repo.
There are commented out print statements, single and two letter
variables, and questionable organization decisions everywhere.

That said, this code works and it's not "hacky". It's based on bounding
boxes and works better than any physics code SpockBot has had to date.
At the moment Spock can cleanly run through drops that would have
caught him with any previous iteration of this plugin.

Note: SpockBot is still "chubby", his width bounding box is 0.2 blocks
wider than a standard player bounding box. While this is probably no
longer necessary, I left it as a "just in case".